Update TipNumber Creator with Wolf RPG Editor.js

+++ Content/TipNumber Creator with Wolf RPG Editor/Script/TipNumber Creator with Wolf RPG Editor.js
@@ -42,6 +42,6 @@ function Create(File) {
 		Reader.readAsArrayBuffer(File);
 		
 		Reader.onload = function () {
-			MapDatas = new Uint32Array(Reader.result);
+			MapDatas = new Uint8Array(Reader.result);
 		}
 }
